feat: updated back-end

+++ server/src/routes/memories.ts
@@ -3,8 +3,17 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
+
+  app.get("/memories", async (request) => {
+    // verifica se está vindo o token de autenticação
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
         // DATAS DA MAIS ANTIGA PRA MAIS NOVA / FORMA CRESCENTE
@@ -15,11 +24,12 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id: memorie.id,
         coverUrl: memorie.coverUrl,
         excerpt: memorie.content.substring(0, 115).concat("..."),
+        createdAt: memorie.createdAt,
       };
     });
   });
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     // const { id } = request.params;
     const paramsSchema = z.object({
       id: z.string().uuid(),
@@ -32,6 +42,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     });
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
     return memory;
     // ENCONTRE A MEMÓRIA COM ESTE ID, SE N ENCONTRAR DISPARE UM ERRO
   });
@@ -52,13 +66,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "52547c7d-137a-415f-ab7d-673c9a116d11",
+        userId: request.user.sub,
       },
     });
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -73,7 +87,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, isPublic, coverUrl } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -86,12 +109,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: {
